Move utility files to new utils directory
