+++ laplace/baselaplace.py
@@ -4,8 +4,8 @@ import torch
 from torch.nn.utils import parameters_to_vector, vector_to_parameters
 from torch.distributions import MultivariateNormal, Dirichlet, Normal
 
-from laplace.utils import parameters_per_layer, invsqrt_precision, get_nll, validate
-from laplace.matrix import Kron
+from laplace.utils.utils import parameters_per_layer, invsqrt_precision, get_nll, validate
+from laplace.utils.matrix import Kron
 from laplace.curvature import BackPackGGN, AsdlHessian
 
 
@@ -754,7 +754,7 @@ class KronLaplace(ParametricLaplace):
     Mathematically, we have for each parameter group, e.g., torch.nn.Module,
     that \\P\\approx Q \\otimes H\\.
     See `BaseLaplace` for the full interface and see
-    `laplace.matrix.Kron` and `laplace.matrix.KronDecomposed` for the structure of
+    `laplace.utils.matrix.Kron` and `laplace.utils.matrix.KronDecomposed` for the structure of
     the Kronecker factors. `Kron` is used to aggregate factors by summing up and
     `KronDecomposed` is used to add the prior, a Hessian factor (e.g. temperature),
     and computing posterior covariances, marginal likelihood, etc.
@@ -812,7 +812,7 @@ class KronLaplace(ParametricLaplace):
 
         Returns
         -------
-        precision : `laplace.matrix.KronDecomposed`
+        precision : `laplace.utils.matrix.KronDecomposed`
         """
         self._check_H_init()
         return self.H * self._H_factor + self.prior_precision
