+++ tests/test_lllaplace.py
@@ -8,8 +8,8 @@ from torch.utils.data import DataLoader, TensorDataset
 from torch.distributions import Normal, Categorical
 from torchvision.models import wide_resnet50_2
 
-from laplace.lllaplace import LLLaplace, FullLLLaplace, KronLLLaplace, DiagLLLaplace
-from laplace.feature_extractor import FeatureExtractor
+from laplace.lllaplace import FullLLLaplace, KronLLLaplace, DiagLLLaplace
+from laplace.utils.feature_extractor import FeatureExtractor
 from tests.utils import jacobians_naive
